backend/requirements.txt
astroid
Click
ecdsa
Flask
Flask-SQLAlchemy
future
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pycryptodome
pylint
python-jose
six
typed-ast
Werkzeug
wrapt
Flask-Cors